backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-magic==0.4.27
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.1
pydantic-settings==2.1.0
python-chess==1.999
stockfish==3.28.0
openai==1.14.3
python-json-logger==2.0.7
